Update
Added static map caching

+++ services/GoogleMaps_StaticMapService.php
@@ -5,6 +5,13 @@ class GoogleMaps_StaticMapService extends BaseApplicationComponent
 {
     public $url = 'https://maps.googleapis.com/maps/api/staticmap';
 
+    protected $expirationLength = false;
+
+    public function __construct()
+    {
+        $this->expirationLength = craft()->config->get('staticMapCacheLength', 'googlemaps');
+    }
+
     public function image(GoogleMaps_StaticMapModel $data, $options = array())
     {
         $url = $this->generate($data, $options);
@@ -14,31 +21,66 @@ class GoogleMaps_StaticMapService extends BaseApplicationComponent
 
     public function generate(GoogleMaps_StaticMapModel $data, $options = array())
     {
-        return $this->url . '?' . $data->getParameters();
+        $basePath = craft()->config->get('staticMapCachePath', 'googlemaps');
+        $baseUrl = craft()->config->get('staticMapCacheUrl', 'googlemaps');
 
-        $client = new \Guzzle\Http\Client;
-        
-        $response = $client->get($this->url . '?' . $data->getParameters());
+        $url = $this->url . '?' . $data->getParameters();
 
-        $response->send();
-
-        $rawdata = (string) $response->getResponse()->getBody();
-
-        $storagePath = craft()->path->getStoragePath() . 'googlemaps/static/';
-
-        IOHelper::ensureFolderExists($storagePath);
-
-        $filename = md5($storagePath . time()) . '.' . $data->format;
-        $fullpath = $storagePath . $filename;
-
-        if(file_exists($fullpath))
+        if($this->expirationLength)
         {
-            unlink($fullpath);
+            $expires = date('Y-m-d H:i:s', time() - $this->expirationLength * 24 * 60 * 60);
+        }
+        else
+        {
+            $expires = '0000-00-00 00:00:00';
         }
 
-        $fp = fopen($fullpath,'x');
-        fwrite($fp, $rawdata);
-        fclose($fp);
+        $record = GoogleMaps_StaticMapRecord::model()->find('query = :query AND dateCreated >= :date', array(
+            ':query' => $data->getParameters(),
+            ':date' => $expires
+        ));
+
+        if($record && $record->cachedFileExists())
+        {
+            return $record->getCachedUrl();
+        }
+
+        if($basePath && $baseUrl)
+        {
+            $basePath = rtrim($basePath, '/') . '/';
+            $baseUrl = rtrim($baseUrl, '/') . '/';
+
+            $client = new \Guzzle\Http\Client;
+            
+            $response = $client->get($url);
+
+            $response->send();
+
+            $rawdata = (string) $response->getResponse()->getBody();
+
+            IOHelper::ensureFolderExists($basePath);
+
+            $filename = md5($basePath . time()) . '.' . $data->format;
+            $fullpath = $basePath . $filename;
+
+            if(file_exists($fullpath))
+            {
+                unlink($fullpath);
+            }
+
+            $fp = fopen($fullpath,'x');
+            fwrite($fp, $rawdata);
+            fclose($fp);
+
+            $record = new GoogleMaps_StaticMapRecord;
+            $record->query = $data->getParameters();
+            $record->filename = $filename;
+            $record->save();
+
+            return $baseUrl . $filename;
+        }
+
+        return $url;
     }
 
 }
